added a service function to get movie details and did some reorginazing
